% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bucket_path}
\alias{bucket_path}
\title{Analysis-runner GCS bucket path}
\usage{
bucket_path(path, bucket_category = NULL)
}
\arguments{
\item{path}{A path to append to the bucket.}

\item{bucket_category}{A category like "upload", "tmp", "web". If omitted,
defaults to the "main" and "test" buckets based on the access level. See
https://github.com/populationgenomics/team-docs/tree/main/storage_policies
for a full list of categories and their use cases.}
}
\value{
Full GCS path.
}
\description{
Returns a full GCS path for the given bucket category and path.
This is useful for specifying input files, as in contrast to the
\code{\link{output_path}} function, \code{bucket_path} does \emph{not} take the "output"
parameter from the analysis-runner invocation into account.
Requires the \code{DATASET} and \code{ACCESS_LEVEL} environment variables to be set.
}
\examples{
# Assuming that the analysis-runner has been invoked with
# `--dataset tob-wgs --access-level test --output snp/v1`:
Sys.setenv("DATASET" = "proj1", "ACCESS_LEVEL" = "test", "OUTPUT" = "dirA/v1")
bucket_path("dir/v1/file.txt") # gs://cpg-proj1-test/dir/v1/file.txt
bucket_path("dir/v1/report.html", "web") # gs://cpg-proj1-test-web/dir/v1/report.html

}
\seealso{
\code{\link{output_path}}
}
