% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{output_path}
\alias{output_path}
\title{Analysis-runner GCS output path}
\usage{
output_path(path_suffix, bucket_category = NULL)
}
\arguments{
\item{path_suffix}{A suffix to append to the bucket + output directory.}

\item{bucket_category}{A category like "upload", "tmp", "web". If omitted,
defaults to the "main" and "test" buckets based on the access level. See
https://github.com/populationgenomics/team-docs/tree/main/storage_policies
for a full list of categories and their use cases.}
}
\value{
Full GCS path.
}
\description{
In contrast to the \code{\link{bucket_path}} function, \code{output_path} takes
the "output" parameter from the analysis-runner invocation into account.
Requires the \code{DATASET}, \code{ACCESS_LEVEL}, and \code{OUTPUT} environment variables to
be set.
}
\examples{
# Assuming that the analysis-runner has been invoked with
# `--dataset tob-wgs --access-level test --output snp/v1`:
Sys.setenv("DATASET" = "proj1", "ACCESS_LEVEL" = "test", "OUTPUT" = "dirA/v1")
(o1 <- output_path("report.html", "web")) # gs://cpg-proj1-test-web/dirA/v1/report.html
(o2 <- output_path("output.txt")) # gs://cpg-proj1-test/dirA/v1/output.txt

}
\seealso{
\code{\link{bucket_path}}
}
